requirements.txt
seleniumbase
beautifulsoup4
python-dotenv